migrating key functions to app controller
