+++ service/controller/appcatalog/v1/key/key.go
@@ -7,34 +7,6 @@ import (
 	"github.com/giantswarm/app-operator/pkg/label"
 )
 
-func AppCatalogTitle(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Title
-}
-
-func AppCatalogStorageURL(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Storage.URL
-}
-
-func ConfigMapName(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Config.ConfigMap.Name
-}
-
-func ConfigMapNamespace(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Config.ConfigMap.Namespace
-}
-
-func IsDeleted(cr v1alpha1.AppCatalog) bool {
-	return cr.DeletionTimestamp != nil
-}
-
-func SecretName(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Config.Secret.Name
-}
-
-func SecretNamespace(customResource v1alpha1.AppCatalog) string {
-	return customResource.Spec.Config.Secret.Namespace
-}
-
 // ToCustomResource converts value to v1alpha1.AppCatalog and returns it or error
 // if type does not match.
 func ToCustomResource(v interface{}) (v1alpha1.AppCatalog, error) {
