+++ service/controller/appcatalog/v1/key/key_test.go
@@ -8,119 +8,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func Test_AppCatalogStorageURL(t *testing.T) {
-	expectedURL := "http://giantswarm.io/sample-catalog/"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Title:       "giant-swarm-title",
-			Description: "giant-swarm app catalog sample",
-			Storage: v1alpha1.AppCatalogSpecStorage{
-				Type: "helm",
-				URL:  "http://giantswarm.io/sample-catalog/",
-			},
-		},
-	}
-
-	if AppCatalogStorageURL(obj) != expectedURL {
-		t.Fatalf("app catalog storage url %s, want %s", AppCatalogStorageURL(obj), expectedURL)
-	}
-}
-
-func Test_AppCatalogTitle(t *testing.T) {
-	expectedName := "giant-swarm-title"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Title:       "giant-swarm-title",
-			Description: "giant-swarm app catalog sample",
-			Storage: v1alpha1.AppCatalogSpecStorage{
-				Type: "helm",
-				URL:  "http://giantswarm.io/sample-catalog.tgz",
-			},
-		},
-	}
-
-	if AppCatalogTitle(obj) != expectedName {
-		t.Fatalf("app catalog name %s, want %s", AppCatalogTitle(obj), expectedName)
-	}
-}
-func Test_ConfigMapName(t *testing.T) {
-	expectedName := "giant-swarm-configmap-name"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Config: v1alpha1.AppCatalogSpecConfig{
-				ConfigMap: v1alpha1.AppCatalogSpecConfigConfigMap{
-					Name:      "giant-swarm-configmap-name",
-					Namespace: "giant-swarm-configmap-namespace",
-				},
-			},
-		},
-	}
-
-	if ConfigMapName(obj) != expectedName {
-		t.Fatalf("ConfigMapName %#q, want %#q", ConfigMapName(obj), expectedName)
-	}
-}
-
-func Test_ConfigMapNamespace(t *testing.T) {
-	expectedNamespace := "giant-swarm-configmap-namespace"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Config: v1alpha1.AppCatalogSpecConfig{
-				ConfigMap: v1alpha1.AppCatalogSpecConfigConfigMap{
-					Name:      "giant-swarm-configmap-name",
-					Namespace: "giant-swarm-configmap-namespace",
-				},
-			},
-		},
-	}
-
-	if ConfigMapNamespace(obj) != expectedNamespace {
-		t.Fatalf("ConfigMapNamespace %#q, want %#q", ConfigMapNamespace(obj), expectedNamespace)
-	}
-}
-
-func Test_SecretName(t *testing.T) {
-	expectedName := "giant-swarm-secret-name"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Config: v1alpha1.AppCatalogSpecConfig{
-				Secret: v1alpha1.AppCatalogSpecConfigSecret{
-					Name:      "giant-swarm-secret-name",
-					Namespace: "giant-swarm-secret-namespace",
-				},
-			},
-		},
-	}
-
-	if SecretName(obj) != expectedName {
-		t.Fatalf("SecretName %#q, want %#q", SecretName(obj), expectedName)
-	}
-}
-
-func Test_SecretNamespace(t *testing.T) {
-	expectedNamespace := "giant-swarm-secret-namespace"
-
-	obj := v1alpha1.AppCatalog{
-		Spec: v1alpha1.AppCatalogSpec{
-			Config: v1alpha1.AppCatalogSpecConfig{
-				Secret: v1alpha1.AppCatalogSpecConfigSecret{
-					Name:      "giant-swarm-secret-name",
-					Namespace: "giant-swarm-secret-namespace",
-				},
-			},
-		},
-	}
-
-	if SecretNamespace(obj) != expectedNamespace {
-		t.Fatalf("SecretNamespace %#q, want %#q", SecretNamespace(obj), expectedNamespace)
-	}
-}
-
 func Test_ToCustomResource(t *testing.T) {
 	testCases := []struct {
 		name           string
